requirements.txt
pandas==2.2.2
matplotlib==3.7.5
scikit-learn==1.2.2
tensorflow==2.16.1
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
numpy>=1.22.0  
pillow>=10.0.0
requests>=2.31.0
pytest>=8.0.0
httpx>=0.27.0 